+++ src/shell/CMakeLists.txt
@@ -8,7 +8,7 @@ set(MY_PROJ_SRC "linenoise/linenoise.c" "sds/sds.c")
 # Search mode for source files under CURRENT project directory?
 # "GLOB_RECURSE" for recursive search
 # "GLOB" for non-recursive search
-set(MY_SRC_SEARCH_MODE "GLOB")
+set(MY_SRC_SEARCH_MODE "GLOB_RECURSE")
 
 set(MY_PROJ_INC_PATH
     "${PEGASUS_PROJECT_DIR}/src/include"
