+++ src/shell/commands/gloable_properties.cpp
@@ -1,0 +1,66 @@
+#include "shell/commands.h"
+
+bool use_app_as_current(command_executor *e, shell_context *sc, arguments args)
+{
+    if (args.argc == 1) {
+        if (sc->current_app_name.empty()) {
+            fprintf(stderr, "Current app not specified.\n");
+            return false;
+        } else {
+            fprintf(stderr, "Current app: %s\n", sc->current_app_name.c_str());
+            return true;
+        }
+    } else if (args.argc == 2) {
+        sc->current_app_name = args.argv[1];
+        fprintf(stderr, "OK\n");
+        return true;
+    } else {
+        return false;
+    }
+}
+
+bool process_escape_all(command_executor *e, shell_context *sc, arguments args)
+{
+    if (args.argc == 1) {
+        fprintf(stderr, "Current escape_all: %s.\n", sc->escape_all ? "true" : "false");
+        return true;
+    } else if (args.argc == 2) {
+        if (strcmp(args.argv[1], "true") == 0) {
+            sc->escape_all = true;
+            fprintf(stderr, "OK\n");
+            return true;
+        } else if (strcmp(args.argv[1], "false") == 0) {
+            sc->escape_all = false;
+            fprintf(stderr, "OK\n");
+            return true;
+        } else {
+            fprintf(stderr, "ERROR: invalid parameter.\n");
+            return false;
+        }
+    } else {
+        return false;
+    }
+}
+
+bool process_timeout(command_executor *e, shell_context *sc, arguments args)
+{
+    if (args.argc == 1) {
+        fprintf(stderr, "Current timeout: %d ms.\n", sc->timeout_ms);
+        return true;
+    } else if (args.argc == 2) {
+        int timeout;
+        if (!dsn::buf2int32(args.argv[1], timeout)) {
+            fprintf(stderr, "ERROR: parse %s as timeout failed\n", args.argv[1]);
+            return false;
+        }
+        if (timeout <= 0) {
+            fprintf(stderr, "ERROR: invalid timeout %s\n", args.argv[1]);
+            return false;
+        }
+        sc->timeout_ms = timeout;
+        fprintf(stderr, "OK\n");
+        return true;
+    } else {
+        return false;
+    }
+}
